+++ shared/store/configure-sagas.tsx
@@ -1,6 +1,7 @@
 import chat2Saga from '../actions/chat2'
 import configSaga from '../actions/config'
 import createSagaMiddleware from 'redux-saga'
+import deeplinksSaga from '../actions/deeplinks'
 import deviceSaga from '../actions/devices'
 import fsSaga from '../actions/fs'
 import gitSaga from '../actions/git'
@@ -27,6 +28,7 @@ import * as Saga from '../util/saga'
 function* mainSaga(): Saga.SagaGenerator<any, any> {
   yield Saga.spawn(chat2Saga)
   yield Saga.spawn(configSaga)
+  yield Saga.spawn(deeplinksSaga)
   yield Saga.spawn(deviceSaga)
   yield Saga.spawn(fsSaga)
   yield Saga.spawn(gregorSaga)
